requirement.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Flask==3.1.0
GDAL==3.11.0
geos==0.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.2.5
pillow==11.2.1
psycopg2==2.9.10
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.13.2
Werkzeug==3.1.3
